split test stages
